+++ src/routes/index.js
@@ -1,3 +1,3 @@
-export { default as AppRoutes } from "./AppRoutes";
-export { default as PrivateRoute } from "./PrivateRoute";
-export { default as PublicRoute } from "./PublicRoute";
+export * from "./PrivateRoute";
+export * from "./PublicRoute";
+export * from "./AppRoutes";
